requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dataclasses-json==0.6.7
ddgs==9.6.1
duckduckgo_search==8.1.1
exceptiongroup==1.3.0
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.7.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.37
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.8.0
numpy==2.2.6
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.44
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
zstandard==0.25.0
